feat: enabling cjs and esm exports
