+++ src/lib/index.ts
@@ -1,0 +1,1 @@
+export const foo = 'bar'
