+++ tsconfig.esm.json
@@ -2,9 +2,10 @@
 	"extends": "./tsconfig.json",
 	"compilerOptions": {
 		"rootDir": "./src",
-		"outDir": "./dist",
-		"module": "commonjs",
-		"target": "es2016"
+		"outDir": "./dist/esm",
+		"moduleResolution": "node",
+		"module": "esnext",
+		"target": "esnext"
 	},
 	"exclude": ["tests"]
 }
